Changes in Add Customer Form (#13)
* Changes in Add Customer Form

* Changes in customer form
