+++ frontend/src/App.js
@@ -1,13 +1,13 @@
-import './App.css'
-import React from 'react';
-import Config from './Components/SideBar/Config';
-import LoginPage from './Pages/LoginPage/index';
 
+import React from "react";
+import "../node_modules/bootstrap/dist/css/bootstrap.min.css";
+import CustomerPopUp from "./Components/PopCustomer/CustomerPopUp";
 
 function App() {
   return (
     <div className="App">
-      <Config/>
+      {/* <LoginPage/> */}
+      <CustomerPopUp />
     </div>
   );
 }
