+++ frontend/src/Components/PopCustomer/CustomerPopUp.css
@@ -1,0 +1,16 @@
+.input-style {
+  margin-bottom: 5%;
+}
+.modal-content {
+  width: 125%;
+}
+
+.form-style {
+  margin: 4%;
+}
+
+.button-style {
+  background-color: #16c5d5;
+  color: white;
+  border: none;
+}
